[WIP] alternative processing
Process array by columns first, hen insert all the corners.
Issue: if we want to add the edges in Free/Used transitions, then columns do not
have the same size, as it depends on the numnber of pieces by column.

Extending every column when adding an edge in one has the same complexity as the
previous method (looking at each column and insert the proper char at the right
position).

Putting this method in pause, as I will try to modify the underlying array to:
- add the transition first with a custom number / Slot type
- change the offset list to represent no longer the rows, but the 'between' of
  each row. Thus the offset_list will have a length of `nrow() + 1` (0 + end +
  everything else)