puzzle: switch to 'dividing' offset
Previously, each array cell was either a start or a stop (or a filling cell, if
another piece was starting/ending at this address, located in the middle of the
piece). Meaning that two pieces, one starting at address 'a', and the other
ending at the same address 'a', were filling a cell in the same row. When
displaying them, it may lead to confusion, as the two pieces were somewhat
concurrent at the same address.

Switching to a new model, where a cell is not defined by an address anymore, but
by two: one before, and one after.

Previously:
| 1  | <- Address 'a'
|----|
| 2  | <- Address 'b'

Now:
|----| <- Address 'a'
| 1  |
|----| <- Address 'c'
| 2  |
|----| <- Address 'b'

Now '1' is defined to be "before address 'c'" and "after address 'a'", instead
of "beeing at address 'a'"

The list of addresses/offset has thus no longer the same size as the array
number of rows.
Inserting a row is now an easier step.