Use multi-threading to find each binaries
Share the FlashImage between all threads, as well as the PuzzleDisplay. Since
the latter is modified when a new 'piece' is added, make sure (via Mutex<>) that
it is mutable & thread safe.

Use of Arc<> to share variables between threads, such as the two quoted aboce,
as well as the binary name.